+++ src/basic/constants/index.ts
@@ -1,1 +1,2 @@
 export * from './product';
+export * from './coupon';
